requirements.txt
pandas==2.1.3
openpyxl==3.1.2
matplotlib==3.8.2
gTTS==2.5.0
Pillow==10.1.0
numpy==1.26.2
requests==2.31.0
